Fix Int128.of_string (and maybe other functions)
Int128.of_string was failing with invalid_argument because the comparison of
[n] with [thresh] in lib/str_conv.ml:of_string was returning wrong results
(as if 0 was greater than [thresh]).

This was because OCaml compare function, on custom blocks, check that the
compare functions of both compared values are the same ; if not, it compares
the identifier instead:

    case Custom_tag: {
      int res;
      int (*compare)(value v1, value v2) = Custom_ops_val(v1)->compare;
      /* Hardening against comparisons between different types */
      if (compare != Custom_ops_val(v2)->compare) {
        return strcmp(Custom_ops_val(v1)->identifier,
                      Custom_ops_val(v2)->identifier) < 0
               ? LESS : GREATER;
      }

and many Int128 operations actually return an uint128 custom block (different
compare function).

So now that the identifiers are different (one is "int128" and the other
"uint128" the int128 will always be smaller than the uint128. Before that
change, all int128 would always compare equal to all uint128, which would
still trigger the [of_string] bug.

The solution is to never return an uint128 custom block as a result of a
int128 operation. This patch therefore replaces the uint128 operations
that were used for int128s by the equivalent ones returning proper int128.

Which is enough to fix the tests.