Fix of_string for negative strings
Of the two tests for overflow only the second one was actually designed to
work with negative numbers.
Also, the first test was just comparing n with max/base while the second
was comparing n*base+d with d, assuming all integer implementation would
wrap around after max_int, which I found during tests not to be the case.

So I replaced the two tests by, in essence:

  If n > max/base || n = max/base && d > max%base then fail

just using compare to avoid comparing both for > and then for =,
and with < and min instead of > and max for negative values.