Store 128-bit numbers twice in OCaml values
Deals with 128-bit alignment. The value may move both if it is promoted to
the major heap and subsequently by the compactor and 16-byte alignment is
not guaranteed.

This approach allocated a 5 word custom value and stores the 128 bit number
in words 1+2 and 4+5. It will always be the case that one is aligned and one
is not.